Add files via upload
Only has unfinished schematic for now. All op-amps/diodes/switches/resistors that (I think) should be there are there, but resistor values need to be figured out. There is also already an RP2040 and a ADC (feel free to change the ADC) but there are no connections between them and any other part of the circuit